license.html: Add a blurb about copyright ownership
Software Carpentry doesn't use copyright assignment, so the SWC
material is not actually copyright SWC.  The original content authors
retain copyright and license the material to SWC, which in turn is
relicensing it to whoever is reading license.html or cloning a Git
repository.

This means that (under the CC BY and MIT licenses) we need to be
telling these people who owns the copyright to the content we're
relicensing.  Since SWC produces a number of mostly-distinct
repositories, it seems like a good idea to distribute this information
in repository-specific `AUTHORS` files.