bootcamps/index.html: Different markers for restricted boot camps
I made the marker legend a paragraph, since I was concerned about a
wide table if we went with:

  ? Future Open Boot Camps  ? Future Restricted Boot Camps  ...

and a grid of future/past open/restricted seemed too confusing ;).
The paragraph should flow reasonably well, even on narrow screens.

I also removed the in-tree versions of the old pins, since I haven't
checked the legality of that [1], and if we can't get to
chart.apis.google.com to fetch the markers, we're probably not going
to get the map from maps.googleapis.com either ;).

If the number of markers continues to grow, it may be worth bundling
them into some sort of structure to avoid duplicate entries in the
JavaScript header and the in-body legend, but I don't think we're
quite there yet.

[1]: https://github.com/swcarpentry/website/pull/192#issuecomment-11928689