Adding a Staffing page showing who's teaching where.
The staffing is generated by hand from another (private) repo and will have
to be updated manually from time to time by someone with access to that repo.
This commit also includes a couple minor style changes in bootcamps/index.html.
Most of this work was originally done by Greg Wilson, I just cleaned it up
for a clean merge.