cumulative-enrollment.png: UK to US spelling "enrolment" -> "enrollment"
On Mon, Mar 18, 2013 at 04:03:04AM -0700, Greg Wilson wrote:
> On 2013-03-17 10:44 PM, W. Trevor King wrote:
> > On Sun, Mar 17, 2013 at 07:18:46PM -0700, Greg Wilson wrote:
> > > A blog/2013/03/cumulative-enrolment.html (13)
> > > A files/2013/03/cumulative-enrolment.png (0)
> >
> > s/enrolment/enrollment/ for US spelling? Do we have a standardized
> > (-ised?) spelling for SWC?
>
> Canadians are allowed to use either British or American spelling. I try
> to stick to US, but this one slipped through --- go ahead and merge a fix.

There are a few other "enrolment"s in the boot camp pages, but they're
either for boot camps in the UK (where the UK spelling makes more
sense), or for the GEOMAR bootcamp in Germany (which is closer to the
UK ;), so I left them as-is.

Once we convert to Pelican, we can add UK translations explicitly, and
then everyone will be happy ;).